backend/requirements.txt
fastapi[standard]==0.133.1
PyJWT==2.11.0
pwdlib[argon2]==0.3.0
supabase==2.28.0
python-dotenv==1.2.1
email-validator==2.3.0
python-jobspy==1.1.82
pandas>=2.0.0